Patches and updates
Replace the built-in function provided by the user script manager, GM_download, with GM_xmlhttpRequest. That I was unaware of, This modification simplifies the code and does deal with the requirement for manual download management.

Before starting the download, the "Download All Images" option now makes sure that every image has properly loaded. With this enhancement, the downloaded images are guaranteed to be whole and not cut off or downloaded in part. However, because it waits for each image to load before moving on to the next, this adjustment might cause somewhat slower download speeds. Overall nonetheless, a worthwhile exchange.