fixed some bugs.
file div didn't work as intended it basically just rejects it. since i forgot to set the items to thumb and left out the gallery wip then for the remove switched to the old version since if i use the new version it removes all of the images. so i updated it abit.